+++ packages/icon/src/icons/image-default.tsx
@@ -2,13 +2,8 @@ import { createIcon } from "../create-icon"
 
 export const ImageDefaultIcon = createIcon({
   title: "ImageDefaultIcon",
-  path:
-    <g fillRule="evenodd">
-      <path fill="#FFF" d="M-78-142H922v842H-78z" />
-      <path fillOpacity="0.04" fill="#000" d="M-38-38H62V62H-38z" />
-      <path fill="none" d="M0 0h24v24H0z" />
-      <path
-        d="M19 3c1.103 0 2 .897 2 2v14c0 1.103-.897 2-2 2H5c-1.103 0-2-.897-2-2V5c0-1.103.897-2 2-2zm0 2H5v14h14.002L19 5zm-6 5 5 7H6l3-4 1 1 3-4z"
-        fillOpacity="0.12" fill="#000" fillRule="nonzero" />
-    </g>,
+  viewBox: "0 0 24 24",
+  path: <path fillRule="evenodd" clipRule="evenodd"
+              d="M19 3H5c-1.103 0-2 .897-2 2v14c0 1.103.897 2 2 2h14c1.103 0 2-.897 2-2V5c0-1.103-.897-2-2-2zM5 19V5h14l.002 14H5zm4-6 1 1 3-4 5 7H6l3-4z"
+              fill="currentColor" />,
 })