+++ packages/trigger/src/interface.tsx
@@ -1,9 +1,8 @@
-import { HTMLAttributes, ReactNode } from "react"
+import { ReactNode } from "react"
 
 export type TriggerColorScheme =
   string
   | "white"
-  | "blackAlpha"
   | "gray"
   | "grayBlue"
   | "red"
@@ -28,8 +27,11 @@ export type TriggerPosition =
   | "rt"
   | "rb"
 
-export interface TriggerProps extends HTMLAttributes<HTMLSpanElement> {
+export type TriggerTrigger = "hover" | "click" | "focus"
+
+export interface TriggerProps {
   colorScheme?: TriggerColorScheme
+  trigger?: TriggerTrigger
   content?: string | ReactNode
   position?: TriggerPosition
   showArrow?: boolean
@@ -40,5 +42,6 @@ export interface TriggerProps extends HTMLAttributes<HTMLSpanElement> {
   defaultPopupVisible?: boolean
   popupVisible?: boolean
   disabled?: boolean
+  hasCloseIcon?: boolean
   onVisibleChange?: (visible: boolean) => void
 }