+++ packages/typography/tests/text.real.test.tsx
@@ -39,18 +39,20 @@ test("Text renders with different copy icon", async () => {
 })
 
 test("Text renders with different copy tooltips", async () => {
+  const onCopy = jest.fn()
   render(<Typography>
-    <Text data-testid="test-text" fontSize="20px"
-          copyable={new CopyableBuilder()
-            .copyTooltip("CopyTooltip")
-            .copiedTooltip("CopiedTooltip")
-            .create()}>Text</Text>
+    <Text copyable={new CopyableBuilder()
+      .copyTooltip("CopyTooltip")
+      .copiedTooltip("CopiedTooltip")
+      .onCopy(onCopy)
+      .create()}>Text</Text>
   </Typography>)
   fireEvent.mouseEnter(screen.getByTitle("CopyIcon"))
   await waitFor(() => expect(screen.getByText("CopyTooltip")).toBeInTheDocument(), {
     timeout: 3000,
   })
   fireEvent.click(screen.getByTitle("CopyIcon"))
+  await waitFor(() => expect(onCopy).toBeCalled())
   fireEvent.mouseLeave(screen.getByTitle("CopyIcon"))
   fireEvent.mouseEnter(screen.getByTitle("CopyIcon"))
   await waitFor(() => expect(screen.getByText("CopiedTooltip")).toBeInTheDocument(), {