+++ scripts/prebuild.mjs
@@ -1,4 +1,0 @@
-#!/usr/bin/env zx
-await $`npx lerna bootstrap`
-await $`yarn build-all-components`
-await $`yarn test`