+++ packages/react/CHANGELOG.md
@@ -15,3 +15,4 @@
   - @illa-design/system@0.0.1-alpha.0
   - @illa-design/tag@0.0.1-alpha.0
   - @illa-design/theme@0.0.1-alpha.0
+  - @illa-design/radio@0.0.1-alpha.0